actualizo

+++ server/src/models/projects.js
@@ -17,7 +17,7 @@ sequelize.define('Projects', {
         allowNull: false,
     },
     price: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(11,2),
         allowNull: false,
     },
     visibility: {
